+++ src/Snippets.js
@@ -1,11 +1,14 @@
-define(function (require, exports) {
+define(function (require, exports, module) {
     "use strict";
 
     // Brackets modules
-    var _ = brackets.getModule("thirdparty/lodash");
+    var _               = brackets.getModule("thirdparty/lodash"),
+        ExtensionUtils  = brackets.getModule("utils/ExtensionUtils"),
+        FileSystem      = brackets.getModule("filesystem/FileSystem");
 
     // Local modules
-    var Preferences   = require("src/Preferences"),
+    var ErrorHandler  = require("src/ErrorHandler"),
+        Preferences   = require("src/Preferences"),
         SnippetDialog = require("src/SnippetDialog"),
         Strings       = require("strings"),
         Utils         = require("src/Utils");
@@ -23,6 +26,49 @@ define(function (require, exports) {
     }
 
     function loadSnippet(snippet) {
+        // snippet.source === "directory"
+        // snippet.source === "user";
+        // snippet.source === "gist";
+
+        var ignoreLoad = false;
+
+        var existingSnippet = _.find(SnippetCollection, function (s) {
+            return s.name === snippet.name;
+        });
+
+        if (existingSnippet) {
+            if (!existingSnippet.source || existingSnippet.source === "directory") {
+                // directory snippets can be always overridden
+                ignoreLoad = false;
+            } else if (existingSnippet.source === "user") {
+                // user snippets can only be overriden by user snippets
+                if (snippet.source !== "user") {
+                    ignoreLoad = true;
+                }
+            } else if (existingSnippet.source === "gist") {
+                // gist snippets can be overriden by user snippets and gist snippets
+                if (snippet.source !== "user" && snippet.source !== "gist") {
+                    ignoreLoad = true;
+                }
+            } else {
+                ErrorHandler.show("loadSnippet(): Unknown snippet source: " + existingSnippet.source);
+            }
+        }
+
+        if (ignoreLoad) {
+            console.log("[brackets-snippets] ignoring loading of '" + snippet.name +
+                        "' snippet, because snippet with the same name of source '" + existingSnippet.source +
+                        "' is present.");
+            return;
+        }
+
+        if (existingSnippet) {
+            var io = SnippetCollection.indexOf(existingSnippet);
+            if (io !== -1) {
+                SnippetCollection.splice(io, 1);
+            }
+        }
+
         // every snippets needs to have an unique generated ID
         snippet._id = ++lastSnippetId;
         SnippetCollection.push(snippet);
@@ -63,8 +109,106 @@ define(function (require, exports) {
         });
     }
 
+    function _registerSnippetDirectory(directory) {
+        var snippetDirectories = Preferences.get("snippetDirectories");
+
+        var entry = _.find(snippetDirectories, function (e) {
+            return e.fullPath === directory.fullPath;
+        });
+
+        // if doesn't exist, add it to the collection, automatically load new directories
+        if (!entry) {
+            entry = {
+                fullPath: directory.fullPath,
+                autoLoad: true
+            };
+            snippetDirectories.push(entry);
+            Preferences.set("snippetDirectories", snippetDirectories);
+        }
+    }
+
+    function _loadSnippetDirectories() {
+        var snippetDirectories = Preferences.get("snippetDirectories");
+        snippetDirectories.forEach(function (snippetDirectory) {
+
+            // skip directories we don't want to load on startup
+            if (snippetDirectory.autoLoad !== true) {
+                console.log("[brackets-snippets] skipping directory: " + snippetDirectory.fullPath);
+                return;
+            }
+
+            if (!FileSystem.isAbsolutePath(snippetDirectory.fullPath)) {
+                snippetDirectory.autoLoad = false;
+                ErrorHandler.show("Directory is not an absolute path: " + snippetDirectory.fullPath);
+                Preferences.set("snippetDirectories", snippetDirectories);
+                return;
+            }
+
+            FileSystem.resolve(snippetDirectory.fullPath, function (err, directory) {
+                if (err) {
+                    ErrorHandler.show(err);
+                    return;
+                }
+
+                if (directory.isDirectory !== true) {
+                    snippetDirectory.autoLoad = false;
+                    Preferences.set("snippetDirectories", snippetDirectories);
+                    ErrorHandler.show("_loadSnippetDirectories: " + snippetDirectory.fullPath + " is not a directory!");
+                    return;
+                }
+
+                directory.getContents(function (err, directoryContents) {
+                    if (err) {
+                        ErrorHandler.show(err);
+                        return;
+                    }
+                    directoryContents.forEach(function (snippetFile) {
+                        if (!snippetFile.isFile) {
+                            return;
+                        }
+                        snippetFile.read(function (err, content) {
+                            if (err) {
+                                ErrorHandler.show(err);
+                                return;
+                            }
+                            loadSnippet({
+                                name: snippetFile.name,
+                                template: content,
+                                source: "directory"
+                            });
+                        });
+                    });
+                });
+            });
+        });
+    }
+
+    function loadDefaultSnippets() {
+
+        var modulePath = ExtensionUtils.getModulePath(module);
+        FileSystem.resolve(modulePath + "../default_snippets/", function (err, entry) {
+            if (err) {
+                ErrorHandler.show(err);
+                return;
+            }
+            entry.getContents(function (err, contents) {
+                if (err) {
+                    ErrorHandler.show(err);
+                    return;
+                }
+                // register every directory which contains a set of snippets
+                contents.forEach(function (directory) {
+                    _registerSnippetDirectory(directory);
+                });
+                // now load all of them
+                _loadSnippetDirectories();
+            });
+        });
+    }
+
     function init() {
         loadSnippets();
+        loadDefaultSnippets();
     }
 
     function getAll() {
@@ -83,12 +227,14 @@ define(function (require, exports) {
 
     function addNewSnippetDialog(snippet) {
         return SnippetDialog.show(snippet).done(function (newSnippet) {
+            newSnippet.source = "user";
             loadSnippet(newSnippet);
         });
     }
 
     function editSnippetDialog(snippet) {
         return SnippetDialog.show(snippet).done(function (newSnippet) {
+            newSnippet.source = "user";
             updateSnippet(newSnippet);
         });
     }