+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## 0.0.2 (20/11/2014)
+* Added support for default snippets.
+* Snippets can now be loaded from directories automatically, configure your own directories in settings.
+* Three dot mark (...) will be replaced with a cursor after inserting a snippet.
+
 ## 0.0.1 (19/11/2014)
 * Initial release.